Pass the index type when creating a new Reference
Instead of requiring the hard-coding of the index type inside the
GController sub-classes we can pass it as a parameter when creating
the GControllerReference.

The main rationale for this change is that some data types, like
GHashTable, can have a typed index - e.g. a string, or an integer
pointer, or a direct pointer comparison; it makes sense to allow
passing the type of the indices along when creating the reference
so that the GController's logic resides entirely in the base class
while leaving the option to override the reference creation entirely.

Along with the index type we should also pass the GControllerAction,
so that it gets stored inside the GControllerReference; this allows
storing references and querying the action that led to their creation
at any later point in the application logic.