Add a model example
MySimpleModel is a simple example of a "model" object using a GPtrArray
as storage and a GPtrArrayController internally.

As the model operates on the storage, it also notifies the controller of
the changes; user code (in this case, a simple display function) will
monitor the GController::changed signal on the GController provided by
the model class.